+++ contrib/CMakeLists.txt
@@ -7,8 +7,6 @@ list(APPEND CMAKE_PREFIX_PATH "${CMAKE_CURRENT_LIST_DIR}/cmake")
 #endif()
 
 add_subdirectory(pybind11)
-add_subdirectory(FunctionalPlus)
-add_subdirectory(frugally-deep)
 add_subdirectory(fastText)
 add_subdirectory(googletest)
 
@@ -24,12 +22,6 @@ add_subdirectory(tinyxml2)
 add_subdirectory(boost-cmake)
 add_subdirectory(icu-cmake)
 add_subdirectory(tabulate)
-
-set(frugally-deep_INCLUDE_DIR "${CMAKE_CURRENT_SOURCE_DIR}/frugally-deep/include" CACHE INTERNAL "frugally-deep_INCLUDE_DIR")
-
-set(Eigen3_INCLUDE_DIR "${CMAKE_CURRENT_SOURCE_DIR}/eigen" CACHE INTERNAL "Eigen3_INCLUDE_DIR")
-
-set(FunctionalPlus_INCLUDE_DIR "${CMAKE_CURRENT_SOURCE_DIR}/FunctionalPlus/include" CACHE INTERNAL "FunctionalPlus_INCLUDE_DIR")
 
 set(nlohmann_json_INCLUDE_DIR "${CMAKE_CURRENT_SOURCE_DIR}/json/include" CACHE INTERNAL "nlohmann_json_INCLUDE_DIR")
