+++ src/xmorphy/ml/TFDisambiguator.h
@@ -2,8 +2,6 @@
 
 #pragma once
 #include <xmorphy/ml/Embedding.h>
-#include <xmorphy/ml/KerasModel.h>
-#include <xmorphy/ml/KerasMultiModel.h>
 #include <xmorphy/utils/LRUCache.h>
 #include <xmorphy/ml/TensorflowModel.h>
